refactor: update configuration for receivers and remove unused components, add back speedscope receiver
