+++ src/receivers/mod.rs
@@ -1,4 +1,7 @@
 pub mod abstract_receiver;
 pub mod registry;
+pub mod stack_unwinder;
+
 pub mod txt_receiver;
 pub mod prv_breakdown_receiver;
+pub mod speedscope_receiver;
